+++ src/libretro.c
@@ -101,6 +101,7 @@ static const struct retro_variable prefs[] = {
     { PACKAGE_NAME"_rom", "Thomson model; Auto|TO8|TO8D|TO9|TO9+|MO5|MO6|PC128|TO7|TO7/70" },
     { PACKAGE_NAME"_autorun", "Auto run game; disabled|enabled" },
     { PACKAGE_NAME"_autostart_use_game_hash", "Use game hash for autostart; disabled|enabled" },
+    { PACKAGE_NAME"_autostart_message_hint", "Display hint to start a game; enabled|disabled" },
     { PACKAGE_NAME"_vkb_transparency", "Virtual keyboard transparency; 0%|10%|20%|30%|40%|50%|60%|70%|80%|90%" },
     { PACKAGE_NAME"_floppy_write_protect", "Floppy write protection; enabled|disabled" },
     { PACKAGE_NAME"_tape_write_protect", "Tape write protection; enabled|disabled" },
@@ -460,8 +461,38 @@ static void update_input(void)
   update_input_virtual_keyboard();
 }
 
+static bool is_option_enabled(char* option)
+{
+  struct retro_variable var = {0, 0};
+  var.key = option;
+  if (environ_cb(RETRO_ENVIRONMENT_GET_VARIABLE, &var))
+  {
+    if (strcmp(var.value, "enabled") == 0)
+    {
+      return true;
+    }
+  }
+  return false;
+}
+
+static void display_start_hint()
+{
+  if (is_option_enabled(PACKAGE_NAME"_autostart_message_hint")
+      && !is_option_enabled(PACKAGE_NAME"_autorun"))
+  {
+    unsigned int language = 0;
+    struct retro_message msg = { "Press 'Start' button to start the game", VIDEO_FPS*2 };
+    if (environ_cb(RETRO_ENVIRONMENT_GET_LANGUAGE, &language) && (language == RETRO_LANGUAGE_FRENCH))
+    {
+      msg.msg = "Appuyez sur le bouton 'Start' pour lancer le jeu";
+    }
+    environ_cb(RETRO_ENVIRONMENT_SET_MESSAGE, &msg);
+  }
+}
+
 static void change_model(const char *model)
 {
+  display_start_hint();
   if (strcmp(model, "Auto") == 0)
   {
     // Auto-detection of the model is only done when a game is loaded